+++ demo/js/demo-one.jsx
@@ -1,4 +1,5 @@
 var React = require('react');
+var ReactDOM = require('react-dom');
 var AriaModal = require('../../');
 
 var DemoOne = React.createClass({
@@ -49,4 +50,4 @@ var DemoOne = React.createClass({
   },
 });
 
-React.render(<DemoOne />, document.getElementById('demo-one'));
+ReactDOM.render(<DemoOne />, document.getElementById('demo-one'));