+++ demo/js/demo-two.jsx
@@ -1,4 +1,5 @@
 var React = require('react');
+var ReactDOM = require('react-dom');
 var AriaModal = require('../../');
 
 var DemoTwo = React.createClass({
@@ -66,4 +67,4 @@ var DemoTwo = React.createClass({
   },
 });
 
-React.render(<DemoTwo />, document.getElementById('demo-two'));
+ReactDOM.render(<DemoTwo />, document.getElementById('demo-two'));